Work-in-progress on adding run support
This passes a basic smoke test but needs proper tests and error handling.
