+++ bundle.json
@@ -1,57 +1,17 @@
 {
-    "schemaVersion": "v1.0.0-WD",
     "name": "helloworld",
-    "version": "0.1.2",
-    "description": "An example 'thin' helloworld Cloud-Native Application Bundle",
-    "maintainers": [
-        {
-            "name": "Matt Butcher",
-            "email": "technosophos@gmail.com",
-            "url": "https://example.com"
-        }
-    ],
+    "version": "0.1.1",
     "invocationImages": [
         {
-            "imageType": "docker",
-            "image": "technosophos/helloworld:0.1.0",
-            "digest": "sha256:aaaaaaa..."
-        }
+        "imageType": "docker",
+        "image": "cnab/helloworld:latest"
+      }
     ],
-    "images": [
-        {
-            "image": "technosophos/microservice:1.2.3",
-            "description": "my microservice",
-            "digest": "sha256:aaaaaaaaaaaa...",
-            "uri": "urn:image1uri",
-            "refs": [
-                {
-                    "path": "image1path",
-                    "field": "image.1.field"
-                }
-            ]
-        }
-    ],
+    "images": [],
     "parameters": {
-        "backend_port" : {
-            "type" : "int",
-            "defaultValue": 80,
-            "minValue": 10,
-            "maxValue": 10240,
-            "metadata": {
-               "description": "The port that the back-end will listen on"
-            }
-        }
-    },
-    "credentials": {
-        "kubeconfig": {
-            "path": "/home/.kube/config"
-	},
-        "image_token": {
-            "env": "AZ_IMAGE_TOKEN"
-        },
-        "hostkey": {
-            "path": "/etc/hostkey.txt",
-            "env": "HOST_KEY"
+        "port": {
+            "defaultValue": 8080,
+            "type": "int"
         }
     }
 }
