+++ pyproject.toml
@@ -10,6 +10,7 @@ keywords = ["cnab"]
 
 [tool.poetry.dependencies]
 python = "^3.7"
+docker = "^3.6"
 
 [tool.poetry.dev-dependencies]
 pytest = "^3.4"
